Refresh patches.kabi/0005-kabi-invalidate-removed-field.patch.
Neil Brown has reported a crash caused by the poisoning because
sysfs_notify_dirent checks s_attr.open and dereference it if
non-NULL. s_attr and s_dir are in the same union and
s_dir::children_unused is sharing the offset with sysfs_elem_attr::open
so it would crash when dereferencing poison value.

Drop the poisoning.

suse-commit: 1fcdb56a6877599049ef2e6c9b719027be479f15