Update config files: disable MFD_SYSCON
Disable MFD_SYSCON, GPIO_SYSCON and LEDS_SYSCON in non-arm
configurations. SYSCON is an arm thing, it was enabled accidentally
in other configurations because STMMAC_PLATFORM, which was once
enabled by accident, selects it.

On s390x, as REGMAP_MMIO was also only selected by MFD_SYSCON, we can
disable it too.

suse-commit: d14737d0d0d1b232dda73a4607b51fca2aefb4da