Delete patches.fixes/scsi-ibmvscsi-show-config.patch.
This patch was added to address empty sysfs files in SLES9.  The root
cause was ultimately identified later: the buffer size passed to the
hardware to query was a 16-bit value, so 64k caused nothing to be
returned.  It was fixed correctly in v3.7 commit 225c56960fc
(ibmvscsi: Fix host config length field overflow), rendering this
patch unnecessary.

suse-commit: d1e631702445f1b308e68875dfcc11a764001cb9