- Delete   patches.fixes/x86-allow_kernel_map_pages_in_pgd_to_work_when_nx_is_disabled.patch. - Delete   patches.fixes/x86-configure_nx_support_earlier_in_setup_arch.patch.
These two came from bsc#884369.

Here's the deal: they were supposed to honor the situation where
eXecuteDisable was cleared in the BIOS, i.e., no NX support. But this is
not required and even plain wrong on Linux. So it was fixed differently
upstream:

  04633df0c43d ("x86/cpu: Call verify_cpu() after having entered long mode too")

which we already have in 12SP2 and this makes those patches a NOP.

So remove them.

Signed-off-by: Borislav Petkov <bp@suse.de>

suse-commit: af9594cd2cbc2d8ccecfb9acc6360aad0028bd4d