Delete patches.suse/mm-page_alloc-convert-page_group_by_mobility_disable-to-static-key.patch.
There is one report that shows on an arch we do not support that there
was a warning and is related to available memory. It's highly unlikely
to triggeer iin SLE but the proposed fix is risky. The benefit of the patch
is marginal so it's being deleted due to being the most straight-forward
approach.

suse-commit: c01d867d824b8e3236c2827d3c6beadf914334ff