blacklist.conf:
a04a480d4392 net: Require exact match for TCP socket lookups if dif is l3mdev (v4.9-rc4)
10/11 conflicts are with code introduced by 74b20582ac38 ("net: l3mdev: Add
hook in ip and ipv6", v4.7-rc1) which is not present in SP2. I think that
either the problem was always there or was introduced by 74b20582ac38. If
in the first case, the fix would have to be implemented differently; if in
the second case, the fix is not needed in SP2.

suse-commit: 07d2529301b91d42b0a3e0eab10ef9d93176af55